refactor: Sound.kt
* inline SOUND_PATTERN
* extract SoundSide + SingleSoundSide
* convert 'Sound' to object
