+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/SoundSide.kt
@@ -1,0 +1,25 @@
+/*
+ *  Copyright (c) 2023 David Allison <davidallisongithub@gmail.com>
+ *
+ *  This program is free software; you can redistribute it and/or modify it under
+ *  the terms of the GNU General Public License as published by the Free Software
+ *  Foundation; either version 3 of the License, or (at your option) any later
+ *  version.
+ *
+ *  This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ *  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ *  PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with
+ *  this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.ichi2.anki.cardviewer
+
+/**
+ * The subset of sounds to involve
+ * @param int Used for serialisation
+ */
+enum class SoundSide(val int: Int) {
+    QUESTION(0), ANSWER(1), QUESTION_AND_ANSWER(2);
+}
